requirements.txt
python-telegram-bot==13.3
requests==2.25.0